+++ src/components/AnnouncementsBanner/styles.tsx
