--- src/components/ProposalBanner/index.tsx
@@ -1,1 +1,0 @@
-export * from './ProposalBanner';
